refactor: centralize ManagePromotersPage component
